pyproject.toml
[tool.poetry]
name = "newswather-consumer"
version = "0.1.0"
description = ""
authors = ["david <yilin9999@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.14.0"
kafka-python = "^2.0.1"
pymongo = "^3.10.1"
fire = "^0.3.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
pytest = "^5.4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
